FeatureLinePointsClass
I have added a FeatueLinePoints class. This is used to export featureline-points into a list from a bunch of featurelines with a given code.
In our workflow we do not have much use for the featurelines themselves because we have to create execution models and a such (on linear infro projects) do not have much use for nicly curve walls because we place them in straight segments.
The segments are calculated using a corridor with a fixed frequency and in the export the geometry points of alignments and profiles are ignored.